dacribagents-api/pyproject.toml
[project]
name = "dacribagents"
version = "0.1.0"
description = "Woodcreek Home Agentic AI System - Multi-agent orchestration for home management"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"

dependencies = [
    # === API Framework ===
    "fastapi[standard]>=0.115.8",
    "httpx>=0.27.0",
    # === LangChain / LangGraph Core ===
    "langchain-core>=0.3.34",
    "langchain-community>=0.3.17",
    "langchain-huggingface>=0.1.2",
    "langgraph>=0.2.70",
    # === NVIDIA Ecosystem (Cert-Aligned) ===
    "langchain-nvidia-ai-endpoints>=0.3.0",
    "nemoguardrails>=0.19.0",
    # === Vector Database ===
    "pymilvus>=2.4.0",
    "sentence-transformers>=3.0.0",
    # === State Persistence ===
    "langgraph-checkpoint-postgres>=2.0.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "asyncpg>=0.29.0",
    # === Observability ===
    "langsmith>=0.1.0",
    "honeyhive>=0.2.57",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
    # === Evaluation (Cert-Aligned) ===
    "ragas>=0.1.0",
    # === Structured Outputs ===
    "instructor>=1.0.0",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    # === Utilities ===
    "loguru>=0.7.3",
    "python-dotenv>=1.0.0",
    # === Development ===
    "ipykernel>=6.29.5",
    "pre-commit>=4.1.0",
    "psycopg-binary>=3.3.2",
    "langchain-openai>=1.1.1",
    "tenacity>=9.1.2",
    "boto3>=1.42.11",
    "botocore>=1.42.11",
    "pypdf>=6.4.2",
    "confluent-kafka>=2.3.0",
    "einops>=0.8.1",
    "croniter>=2.0.0",
    "exchangelib>=5.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.9.6",
    "pytest>=8.3.5",
    "pytest-cov",
    "pytest-asyncio>=0.23.0",
    "pyyaml>=6.0.2",
    "poethepoet>=0.33.1",
    "pre-commit>=4.2.0",
    "respx>=0.21.0",          # Mock httpx for testing
    "alembic>=1.13.0",        # Database migrations
    "sqlalchemy>=2.0.0",      # Required by Alembic
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dacribagents"]

# === Ruff Configuration ===
[tool.ruff]
line-length = 119

[tool.ruff.lint]
extend-select = [
    "E",     # pycodestyle errors
    "B",     # flake8-bugbear
    "PL",    # pylint equivalent rules
    "I",     # isort
    "D",     # pydocstyle for docstrings
    "N",     # pep8-naming
    "ASYNC", # flake8-async (important for asyncio code)
]

ignore = [
    "D107",    # Missing docstring in `__init__`
    "E501",    # Line too long
    "D100",    # Missing docstring in public module
    "D104",    # Missing docstring in public package
    "D103",    # Missing docstring in public function
    "D101",    # Missing docstring in public class
    "D401",    # First line of docstring should be in imperative mood
    "PLR2004", # Magic value used in comparison
    "PLW2901", # `with` statement variable overwritten
]

[tool.ruff.lint.mccabe]
max-complexity = 10

# === Pytest Configuration ===
[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]
pythonpath = ["."]
asyncio_mode = "auto"
norecursedirs = [
    "src/dacribagents/samples/python_tests",
    "example-project",
    "tests/artifacts",
]
addopts = [
    "--cov=src",
    "--cov-fail-under=80",
    "--junitxml=test-reports/report.xml",
    "--cov-report=xml",
    "--cov-report=html",
    "--cov-report=term",
]

# === Coverage Configuration ===
[tool.coverage.paths]
source = ["src/"]

[tool.coverage.report]
exclude_also = [
    "pragma: no cover",
    "if False",
    "def __repr__",
    "if self.debug",
    "raise AssertionError",
    "raise NotImplementedError",
    "raise MemoryError",
    "except DistributionNotFound",
    "except ImportError",
    "@abc.abstractmethod",
    "if 0:",
    "if __name__ == .__main__.:",
    "if typing.TYPE_CHECKING:",
    "if TYPE_CHECKING:",
]
omit = ["tests/artifacts/*"]
show_missing = true

[tool.coverage.run]
branch = true
parallel = true
data_file = "test-reports/.coverage"

[tool.coverage.xml]
output = "test-reports/coverage.xml"

[tool.coverage.html]
directory = "test-reports/htmlcov"

[tool.coverage.json]
output = "test-reports/coverage.json"

# === Poe Tasks ===
[tool.poe]
executor.type = "uv"

[tool.poe.tasks.lint]
cmd = "uvx --from pre-commit pre-commit run --all-files"
help = "Run linting, formatting, and static code quality tools"

[tool.poe.tasks.test]
cmd = "pytest"
help = "Run all tests with optional additional arguments"

[tool.poe.tasks.test-fast]
cmd = "pytest -m 'not slow and not integration'"
help = "Run fast unit tests only"

[tool.poe.tasks.serve-coverage-report]
help = "Serve the coverage report on http://localhost:3333"
shell = """
echo "Serving coverage report on http://localhost:3333"
echo "Press Ctrl+C to stop the server"
python -m http.server 3333 --directory ./test-reports/htmlcov
"""

[tool.poe.tasks.dev]
cmd = "fastapi dev src/dacribagents/api/main.py --port 8090"
help = "Run FastAPI development server with hot reload"

[tool.poe.tasks.clean]
help = "Remove all generated files"
shell = """
rm -rf dist build coverage.xml test-reports sample/ tests/artifacts/

find . \\
    -type d \\
    \\( \\
    -name "*cache*" \\
    -o -name "tests/artifacts" \\
    -o -name "*.dist-info" \\
    -o -name "*.egg-info" \\
    -o -name "*htmlcov" \\
    -o -name "*.metaflow" \\
    -o -name "*.metaflow.s3" \\
    -o -name "*.mypy_cache" \\
    -o -name "*.pytest_cache" \\
    -o -name "*.ruff_cache" \\
    -o -name "*__pycache__" \\
    \\) \\
    -not -path "*env/*" \\
    -exec rm -r {} + || true

find . \\
    -type f \\
    -name "*.pyc" \\
    -o -name "*.DS_Store" \\
    -o -name "*.coverage*" \\
    -not -path "*env/*" \\
    -exec rm {} +
"""
